requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.8.0
    # via flask-migrate
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via jsonschema
brotli==1.0.9
    # via flask-compress
cachelib==0.9.0
    # via flask-session
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via cryptography
chardet==4.0.0
    # via prance
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   clickclick
    #   flask
clickclick==20.10.2
    # via connexion
commonmark==0.9.1
    # via rich
connexion==2.14.0
    # via -r requirements.in
cryptography==37.0.2
    # via
    #   msal
    #   pyjwt
cssmin==0.2.0
    # via -r requirements.in
deprecated==1.2.13
    # via redis
dnspython==2.2.1
    # via email-validator
email-validator==1.2.1
    # via -r requirements.in
flask==2.1.1
    # via
    #   -r requirements.in
    #   connexion
    #   flask-assets
    #   flask-compress
    #   flask-migrate
    #   flask-redis
    #   flask-session
    #   flask-sqlalchemy
    #   flask-wtf
    #   funding-service-design-utils
flask-assets==2.0
    # via -r requirements.in
flask-compress==1.10.1
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-redis==0.4.0
    # via -r requirements.in
flask-session==0.3.2
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
flask-talisman==0.8.1
    # via -r requirements.in
flask-wtf==1.0.0
    # via -r requirements.in
funding-service-design-utils @ git+https://github.com/communitiesuk/funding-service-design-utils.git@main
    # via -r requirements.in
govuk-frontend-jinja==2.0.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via funding-service-design-utils
idna==3.3
    # via
    #   email-validator
    #   requests
inflection==0.5.1
    # via connexion
itsdangerous==2.1.2
    # via
    #   connexion
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via
    #   flask
    #   govuk-frontend-jinja
    #   swagger-ui-bundle
jsmin==3.0.1
    # via -r requirements.in
jsonschema==4.6.0
    # via
    #   connexion
    #   openapi-schema-validator
    #   openapi-spec-validator
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
msal==1.18.0
    # via -r requirements.in
openapi-schema-validator==0.2.3
    # via openapi-spec-validator
openapi-spec-validator==0.4.0
    # via -r requirements.in
packaging==21.3
    # via
    #   connexion
    #   redis
prance==0.21.8.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via rich
pyjwt[crypto]==2.4.0
    # via
    #   -r requirements.in
    #   msal
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pyscss==1.4.0
    # via -r requirements.in
python-dotenv==0.20.0
    # via funding-service-design-utils
python-json-logger==2.0.2
    # via funding-service-design-utils
pytz==2022.1
    # via funding-service-design-utils
pyyaml==6.0
    # via
    #   clickclick
    #   connexion
    #   funding-service-design-utils
    #   openapi-spec-validator
redis==4.3.4
    # via flask-redis
requests==2.28.0
    # via
    #   -r requirements.in
    #   connexion
    #   msal
    #   prance
rich==12.4.4
    # via funding-service-design-utils
ruamel-yaml==0.17.21
    # via prance
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
semver==2.13.0
    # via prance
six==1.16.0
    # via
    #   flask-talisman
    #   prance
    #   pyscss
sqlalchemy==1.4.39
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
swagger-ui-bundle==0.0.9
    # via -r requirements.in
urllib3==1.26.9
    # via requests
webassets==2.0
    # via flask-assets
werkzeug==2.1.2
    # via
    #   connexion
    #   flask
wrapt==1.14.1
    # via deprecated
wtforms==3.0.1
    # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
